+++ mirrord-cli/src/config.rs
@@ -14,6 +14,7 @@ pub(super) enum Commands {
         #[clap(value_parser)]
         path: String,
     },
+    Login(LoginArgs),
 }
 
 #[derive(Args, Debug)]
@@ -85,4 +86,23 @@ pub(super) struct ExecArgs {
     /// Enable tcp outgoing feature.
     #[clap(short = 'o', long, value_parser)]
     pub enable_tcp_outgoing: bool,
+}
+
+#[derive(Args, Debug)]
+pub(super) struct LoginArgs {
+    /// Manualy insert token
+    #[clap(long)]
+    pub token: Option<String>,
+
+    /// Time to wait till close the connection wating for reply from identity server
+    #[clap(long, default_value = "120")]
+    pub timeout: u64,
+
+    /// Override identity server url
+    #[clap(long, default_value = "https://identity.metalbear.dev")]
+    pub auth_server: String,
+
+    /// Don't open web browser automatically and just print url
+    #[clap(long)]
+    pub no_open: bool,
 }