+++ CHANGELOG.md
@@ -9,7 +9,16 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ### Added
 - New feature, [tcp outgoing traffic](https://github.com/metalbear-co/mirrord/issues/27). It's now possible to make requests to a remote host from the staging environment context. You can enable this feature setting the `MIRRORD_TCP_OUTGOING` variable to true, or using the `-o` option in mirrord-cli.
+- mirrord-cli add login command for logging in to metalbear-cloud
+
+## 2.9.1
+### Fixed
+- CI - set typescript version at 4.7.4 to fix broken release action
+
+## 2.9.0
+### Added
 - Support for Golang fileops
+- IntelliJ Extension for mirrord
 
 ### Changed
 - mirrord-layer: Added common `Result` type to to reduce boilerplate, removed dependency of `anyhow` crate.